feat: song list API integration with pagination

+++ src/artistMenu/index.jsx
@@ -1,75 +1,68 @@
 import React from 'react';
 import { Pagination } from 'semantic-ui-react';
 import 'semantic-ui-css/semantic.min.css'
+import { connect } from 'react-redux';
 
 import { SongList } from './style';
 import ConfirmationModal from '../common/components/deleteConfirmation';
 import SongForm from '../common/components/songForm';
 import { DarkBackground } from '../common/components/style';
 import Sidebar from '../common/components/sidebar';
+import { getSongs } from './action';
 
 class ArtistMenu extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
-            artist: null,
             loading: true,
             error: null,
             page: 1,
             deleteId: null,
-            editSong: null,
+            editSongId: null,
+            editSongName: ''
         };
     }
 
-    // componentDidMount() {
-    //     this.fetchArtist();
-    // }
-
-    // componentDidUpdate(prevProps) {
-    //     if (this.props.artistId !== prevProps.artistId) {
-    //     this.fetchArtist();
-    //     }
-    // }
-
-    // fetchArtist() {
-    //     this.setState({ loading: true });
-    //     fetch(`/api/artists/${this.props.artistId}`)
-    //     .then(response => response.json())
-    //     .then(artist => this.setState({ artist, loading: false }))
-    //     .catch(error => this.setState({ error, loading: false }));
-    // }
-
-    deleteSongModal = (url) => {
-        this.setState({deleteId: url});
-        document.getElementById("delete-confirmation-background").classList.toggle("show");
-        document.getElementById("delete-confirmation-box").classList.toggle("show");
+    componentDidMount = () => {
+        this.props.getSongsFunction(10,this.state.page)
     }
 
-    editSongModal = (url) => {
-        this.setState({editSong: url});
-        document.getElementById("insert-song-form").classList.toggle("show");
-        document.getElementById("dark-background").classList.toggle("show");
+    componentDidUpdate(_, prevState) {
+        if (this.state.page !== prevState.page) {
+            this.props.getSongsFunction(10,this.state.page)
+        }
+    }
+
+    deleteSongModal = (url) => {
+        this.setState({deleteId: url}, () => {
+            document.getElementById("delete-confirmation-background").classList.toggle("show");
+            document.getElementById("delete-confirmation-box").classList.toggle("show");
+        })
+    }
+
+    editSongModal = (id, title) => {
+        this.setState({editSongId: id, editSongName: title}, () => {
+            document.getElementById("insert-song-form").classList.toggle("show");
+            document.getElementById("dark-background").classList.toggle("show");
+        })
     }
 
     closeEditSongModal = () => {
-        this.setState({editSong: null});
+        this.setState({editSongId: null, editSongName: ''});
         document.getElementById("insert-song-form").classList.toggle("show");
         document.getElementById("dark-background").classList.toggle("show");
     }
 
     render() {
-        const { loading, error } = this.state;
-        // if (loading) {
-        //     return <div>Loading...</div>;
-        // } else if (error) {
-        //     return <div>{error.message}</div>;
-        // } else {
+        if (this.props.loading) {
+            return <div>Loading...</div>;
+        } else {
             return (
                 <div>
                     <Sidebar/>
                     <ConfirmationModal type='delete' acceptfunction={() => this.props.deleteSongFunction}/>
                     <DarkBackground id='dark-background' onClick={this.closeEditSongModal}/>
-                    <SongForm title='Edit Song Details' buttontext='Update Song'/>
+                    <SongForm id='song-form-component' title='Edit Song Details' buttontext='Update Song' value={this.state.editSongName}/>
                     <div style={{  
                         margin: '50px 100px',
                         marginLeft: '20rem',
@@ -82,18 +75,19 @@ class ArtistMenu extends React.Component {
                                 <tr>
                                     <th>#</th>
                                     <th>TITLE</th>
-                                    <th>RELEASED</th>
                                     <th>MANAGE</th>
                                 </tr>
-                                <tr className="content" name="$id">
-                                    <td>1</td>
-                                    <td className='songTitle'>Cukup Siti Khodijah</td>
-                                    <td>10/10/2000</td>
+                                {this.props.songs.premium_song.map((song, index) => (
+                                <tr key={song.id} className="content" name="$id">
+                                    <td>{index+1}</td>
+                                    <td className='songTitle'>{song.title}</td>
                                     <td className="manage">
-                                        <p songid={"0"} style={{marginRight:"1rem"}} onClick={(e) => this.editSongModal(e.target.getAttribute('songid'))}>Edit</p>
-                                        <p songid={"0"} onClick={(e) => this.deleteSongModal(e.target.getAttribute('songid'))}>Delete</p>
+                                        <p style={{marginRight:"1rem", marginBottom: "0"}} onClick={(_) => this.editSongModal(song.id, song.title)}>Edit</p>
+                                        <p onClick={(_) => this.deleteSongModal(song.id, song.title)}>Delete</p>
                                     </td>
                                 </tr>
+                                ))
+                                }
                             </tbody>
                         </SongList>
                         <Pagination
@@ -102,14 +96,27 @@ class ArtistMenu extends React.Component {
                             lastItem={null}
                             pointing
                             secondary
-                            totalPages={10}
+                            totalPages={this.props.songs.maximum_page}
                             onPageChange={(e, { activePage }) => this.setState({ page: activePage })}
                         />
                     </div>
                 </div>
             );
-        // }
+        }
     }
 }
 
-export default ArtistMenu;
+const mapStateToProps = (state) => {
+	return {
+        songs: state.artistMenu.songs,
+        loading: state.artistMenu.loading
+	}
+}
+
+const mapDispatchToProps = (dispatch) => {
+    return {
+        getSongsFunction: (limit, page) => dispatch(getSongs(limit, page))
+    }
+}
+
+export default connect(mapStateToProps,mapDispatchToProps)(ArtistMenu);
